requirements.txt
Flask==1.0.2
Click==7.0
Flask-SQLAlchemy==2.3.2
